+++ xnradmin-model/xnradmin-core/src/main/java/com/xnradmin/core/service/business/order/FarmerOrderRecordService.java
@@ -1,6 +1,5 @@
 package com.xnradmin.core.service.business.order;
 
-import java.io.Serializable;
 import java.text.SimpleDateFormat;
 import java.util.LinkedList;
 import java.util.List;
@@ -10,21 +9,15 @@ import org.apache.commons.logging.LogFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.cntinker.util.StringHelper;
 import com.xnradmin.core.dao.CommonDAO;
 import com.xnradmin.core.dao.CommonJDBCDAO;
-import com.xnradmin.core.dao.business.order.BusinessOrderRecordDAO;
 import com.xnradmin.core.dao.business.order.FarmerOrderRecordDAO;
 import com.xnradmin.core.service.StaffService;
 import com.xnradmin.po.business.BusinessGoods;
-import com.xnradmin.po.business.BusinessOrderGoodsRelation;
-import com.xnradmin.po.business.BusinessOrderRecord;
 import com.xnradmin.po.business.FarmerOrderRecord;
-import com.xnradmin.po.mall.commodity.GoodsAllocationShow;
 import com.xnradmin.po.wx.OutPlan;
 import com.xnradmin.po.wx.connect.Farmer;
 import com.xnradmin.vo.business.FarmerOrderVO;
-import com.xnradmin.vo.business.OutPlanVO;
 
 
 /**
@@ -61,11 +54,11 @@ public class FarmerOrderRecordService {
 	}
 
 	public FarmerOrderRecord findByid(String id) {
-		return dao.findById(Long.valueOf(id));
+		return dao.findById(Integer.valueOf(id));
 	}
 
 	public void del(String id) {
-		FarmerOrderRecord po = this.dao.findById(Long.valueOf(id));
+		FarmerOrderRecord po = this.dao.findById(Integer.valueOf(id));
 		this.dao.delete(po);
 	}
